% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lse.R
\name{get_lse}
\alias{get_lse}
\title{Get LSE}
\usage{
get_lse(X, y, algorithm, rep = 20)
}
\arguments{
\item{X, y, algorithm, rep}{the matrix waiting for QR decom}
}
\value{
LSE: beta
}
\description{
This function compute LSE of beta using the result of QR decompostion
}
