% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_decom.R
\name{GramSchm}
\alias{GramSchm}
\title{GramSchm decompostion}
\usage{
GramSchm(A)
}
\arguments{
\item{A}{the matrix waiting for QR decom}
}
\value{
a list of matrx
}
\description{
This function computes the QR decompostion using Gram-Schmidt algorithm
}
