% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_decom.R
\name{Householder}
\alias{Householder}
\title{Householder decompostion}
\usage{
Householder(A)
}
\arguments{
\item{A}{the matrix waiting for QR decom}
}
\value{
a list of matrx
}
\description{
This function computes the QR decompostion using Householder algorithm
}
