% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_decom.R
\name{U2Q}
\alias{U2Q}
\title{Compute orthognal matrix Q from U}
\usage{
U2Q(U)
}
\arguments{
\item{U}{Compute orthognal matrix Q from U}
}
\value{
matrix Q using householder algorithm
}
\description{
Compute orthognal matrix Q from U
}
