% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recursive_lse.R
\name{updated_lse}
\alias{updated_lse}
\title{Recursive Least Square Estimate}
\usage{
updated_lse(X1, X2, y1, y2, beta)
}
\arguments{
\item{X1, X2, y1, y2, beta1}{the matrix waiting for QR decom}
}
\value{
LSE: beta
}
\description{
This function compute LSE recursively
}
